move to recoverwallet namespace
